// src/app/models/person.ts
export class Person {
    username: string;
    first_name: string;
    second_name?:string;
    third_name?: string;
    first_surname: string;
    second_surname?: string;
    third_surname?: string;
    birthdate: string;
    zone_direction: string;
    street_direction: string;
    document_id: string;
    document_expedition?: string;
    cat_document_type?: string;
    cat_genre?: string;
    marital_status?: string;
}